move posts and popular to public api. keep search behind private
