+++ src/store/store.ts
@@ -2,10 +2,11 @@ import type { Action, ThunkAction } from '@reduxjs/toolkit'
 import { combineSlices, configureStore } from '@reduxjs/toolkit'
 import { setupListeners } from '@reduxjs/toolkit/query'
 import { settingsSlice } from './features/settingsSlice'
-import { redditApi } from './services/redditApi'
+import { privateApi } from './services/privateApi'
+import { publicApi } from './services/publicApi'
 
 // Combine all slices into a single reducer function.
-const rootReducer = combineSlices(settingsSlice, redditApi)
+const rootReducer = combineSlices(settingsSlice, publicApi, privateApi)
 
 // Infer the `RootState` type from the root reducer.
 export type RootState = ReturnType<typeof rootReducer>
@@ -22,7 +23,10 @@ export const makeStore = (preloadedState?: Partial<RootState>) => {
   const store = configureStore({
     reducer: rootReducer,
     middleware: (getDefaultMiddleware) =>
-      getDefaultMiddleware().concat(redditApi.middleware),
+      getDefaultMiddleware().concat(
+        publicApi.middleware,
+        privateApi.middleware
+      ),
     preloadedState
   })
   setupListeners(store.dispatch)
